new map (draft)
I created a new map page. Data are fetched from /feed/geo-new.json (I don‘t know how it is done properly). So you would need to change it there in the future. Also the css stuff is now a little bit chaotic. So if someone could clean that it would be great.